Fix Mitaka Multipathing issues
When Multipathing is enabled, os-brick is very insistent that ALL
backends support multipathing in the way it wants. SolidFire does not,
therefore this patch makes os-brick sane, and if it cannot discover another
portal on a device (SolidFire does not) it will just failback to single
paths, which is OK for SolidFire since we have the SVIP to deal with
any network issues.